Collision detection works now for the objects around us, when we have the room I will implement for the walls additionally.
The changes:

init variable added. It is a temporary solution for transforming the BBox initially. BBox is transformed according to modelMatrix.

After the objects moved, the head moves and checks for collisions.

The method detectCollision changed, because I realised when camera moves to +z or +x direction, for the same movement an object should move to -z or -x direction. This difference changed the signs of the border control.